+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.1.1: TBD
 
 * Fixed `ManagedCommand` to provide access to the `Environment`, among other things.
 * Made `JerseyClient`'s thread pool managed.
+* Improved ease of use for `Duration` and `Size` configuration parameters.
 
 
 v0.1.0: Dec 21 2011